chore: convert EditorToolbar to TypeScript
