+++ packages/forma-36-react-components/package.json
@@ -26,19 +26,17 @@
     "@babel/preset-typescript": "7.1.0",
     "@contentful/forma-36-fcss": "^0.0.13",
     "@contentful/forma-36-tokens": "^0.2.0",
-    "@storybook/addon-actions": "^5.0.0",
-    "@storybook/addon-info": "^5.0.0",
-    "@storybook/addon-knobs": "^5.0.0",
-    "@storybook/addon-notes": "^5.0.0",
-    "@storybook/addon-storysource": "^5.0.0",
-    "@storybook/react": "^5.0.0",
+    "@storybook/addon-actions": "^5.0.1",
+    "@storybook/addon-knobs": "^5.0.1",
+    "@storybook/addon-notes": "^5.0.1",
+    "@storybook/addon-storysource": "^5.0.1",
+    "@storybook/react": "^5.0.1",
     "@types/classnames": "2.2.7",
     "@types/enzyme": "3.1.16",
     "@types/jest": "23.3.13",
     "@types/jest-axe": "2.2.2",
     "@types/react": "16.8.6",
     "@types/react-transition-group": "2.0.15",
-    "@types/storybook__addon-info": "4.1.0",
     "@types/storybook__react": "4.0.1",
     "@typescript-eslint/eslint-plugin": "1.3.0",
     "@typescript-eslint/parser": "1.3.0",
