+++ packages/forma-36-react-components/src/components/EditorToolbar/EditorToolbarButton/EditorToolbarButton.stories.tsx
@@ -2,13 +2,14 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { text, select, boolean } from '@storybook/addon-knobs';
 import { action } from '@storybook/addon-actions';
-import { withInfo } from '@storybook/addon-info';
 
 import EditorToolbarButton from './EditorToolbarButton';
-import { iconName } from './../../Icon/constants.ts';
+import { iconName } from './../../Icon/constants';
 
 storiesOf('Components|EditorToolbar/EditorToolbarButton', module)
-  .addDecorator(withInfo)
+  .addParameters({
+    propTypes: EditorToolbarButton['__docgenInfo'],
+  })
   .add('default', () => (
     <EditorToolbarButton
       icon={select('Icon', Object.keys(iconName), Object.keys(iconName)[0])}
