+++ packages/forma-36-react-components/src/components/EditorToolbar/EditorToolbarDivider/EditorToolbarDivider.stories.tsx
@@ -1,12 +1,13 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { text } from '@storybook/addon-knobs';
-import { withInfo } from '@storybook/addon-info';
 
 import EditorToolbarDivider from './EditorToolbarDivider';
 
 storiesOf('Components|EditorToolbar/EditorToolbarDivider', module)
-  .addDecorator(withInfo)
+  .addParameters({
+    propTypes: EditorToolbarDivider['__docgenInfo'],
+  })
   .add('default', () => (
     <EditorToolbarDivider extraClassNames={text('Extra Class Names', '')}>
       EditorToolbarDivider
