+++ packages/forma-36-react-components/src/components/EditorToolbar/EditorToolbarDivider/EditorToolbarDivider.test.tsx
