+++ packages/forma-36-react-components/src/components/EditorToolbar/EditorToolbarDivider/EditorToolbarDivider.tsx
@@ -1,23 +1,21 @@
-import React from 'react';
-import PropTypes from 'prop-types';
+import React, { Component } from 'react';
 import cn from 'classnames';
 import styles from './EditorToolbarDivider.css';
 
-export class EditorToolbarDivider extends React.Component {
-  static propTypes = {
-    testId: PropTypes.string,
-    extraClassNames: PropTypes.string,
-  };
+export interface EditorToolbarDividerProps {
+  testId?: string;
+  extraClassNames?: string;
+}
 
+export class EditorToolbarDivider extends Component<EditorToolbarDividerProps> {
   static defaultProps = {
     testId: 'cf-editor-toolbar-divider',
-    extraClassNames: undefined,
   };
 
   render() {
     const { extraClassNames, testId, ...otherProps } = this.props;
 
-    const classNames = cn(styles.EditorToolbarDivider, extraClassNames);
+    const classNames = cn(styles['EditorToolbarDivider'], extraClassNames);
 
     return (
       <span data-test-id={testId} className={classNames} {...otherProps} />
