+++ packages/forma-36-react-components/src/components/EditorToolbar/EditorToolbarDivider/index.ts
